moved functionality out of React component into redux store.
